Misc updates

+++ CoreMagick/CoreMagick/Controllers/ImagesController.cs
@@ -1,8 +1,11 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Net;
 using System.Net.Http;
+using System.Security.Cryptography;
+using System.Text;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -14,6 +17,8 @@ namespace CoreMagick.Controllers
     [Route("api/v2/images")]
     public class ImagesController : ControllerBase
     {
+        private const string CachePath = "/data";
+        private const string CacheFileExtension = "base64";
         private const string CommandMagickIdentify = "/usr/bin/identify";
         private const string CommandMagickConvert = "/usr/bin/convert";
         private const string CommandArgsIdentify = "-format \"%m\"";
@@ -41,24 +46,84 @@ namespace CoreMagick.Controllers
         {
             try
             {
-                var path = GenerateTemporaryPath();
                 source = WebUtility.UrlDecode(source);
-                _logger.LogInformation($"Retrieving {source}");
-                await RetrieveFile(source, path);
-                var mime = Identify(path);
-                if (resize && width > 0)
+                var cacheString = $"{source}${resize}${width}";
+                var cacheKey = HashMD5(cacheString);
+                var cacheFile = Path.Combine(CachePath, $"{cacheKey}.{CacheFileExtension}");
+                _logger.LogInformation($"{cacheKey}");
+                if (System.IO.File.Exists(cacheFile) && !force)
                 {
-                    _logger.LogInformation($"Resizing (width {width}): {source}");
-                    Resize(path, width);
+                    _logger.LogInformation($"Retrieving from local cache");  
+                    var output = await System.IO.File.ReadAllTextAsync(cacheFile);
+                    return Ok(output);
                 }
-                var output = Base64Encode(path);
-                DeleteTemporaryFile(path);
-                return Ok($"data:{mime};base64,{output}");
+                else
+                {
+                    _logger.LogInformation($"Retrieving from remote {source}");
+                    var path = GenerateTemporaryPath();
+                    await RetrieveFile(source, path);
+                    var mime = Identify(path);
+                    if (resize && width > 0)
+                    {
+                        _logger.LogInformation($"Resizing (width {width}): {source}");
+                        Resize(path, width);
+                    }
+                    var output = Base64Encode(path);
+                    var fullOutput = $"data:{mime};base64,{output}";
+                    await System.IO.File.WriteAllTextAsync(cacheFile, fullOutput);
+                    DeleteTemporaryFile(path);
+                    return Ok(fullOutput);
+                }
             }
             catch (Exception ex)
             {
                 return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
             }
+        }
+        
+        [HttpPost("encode/batch")]
+        public async Task<IActionResult> BatchEncode(
+            [FromBody] List<string> files,
+            [FromQuery(Name = "resize")] bool resize = false,
+            [FromQuery(Name = "width")] ushort width = 0,
+            [FromQuery(Name = "force")] bool force = false)
+        {
+            foreach (var file in files)
+            {
+                try
+                {
+                    var cacheString = $"{file}${resize}${width}";
+                    var cacheKey = HashMD5(cacheString);
+                    var cacheFile = Path.Combine(CachePath, $"{cacheKey}.{CacheFileExtension}");
+                    _logger.LogInformation($"{cacheKey}");
+                    if (System.IO.File.Exists(cacheFile) && !force)
+                    {
+                        // SKIP
+                    }
+                    else
+                    {
+                        _logger.LogInformation($"[Batch] Retrieving from remote {file}");
+                        var path = GenerateTemporaryPath();
+                        await RetrieveFile(file, path);
+                        var mime = Identify(path);
+                        if (resize && width > 0)
+                        {
+                            _logger.LogInformation($"[Batch] Resizing (width {width}): {file}");
+                            Resize(path, width);
+                        }
+
+                        var output = Base64Encode(path);
+                        var fullOutput = $"data:{mime};base64,{output}";
+                        await System.IO.File.WriteAllTextAsync(cacheFile, fullOutput);
+                        DeleteTemporaryFile(path);
+                    }
+                }
+                catch (Exception ex)
+                {
+                    _logger.LogInformation($"[Batch] {ex.Message}");
+                }
+            }
+            return Ok();
         }
         
         [HttpGet("path")]
@@ -144,6 +209,17 @@ namespace CoreMagick.Controllers
             {
                 throw new Exception(error);
             }
+        }
+       
+        private static string HashMD5(string input)
+        {
+            var md5 = MD5.Create();
+            var inputBytes = Encoding.ASCII.GetBytes(input);
+            var hashBytes = md5.ComputeHash(inputBytes);
+            var sb = new StringBuilder();
+            foreach (var t in hashBytes)
+                sb.Append(t.ToString("X2"));
+            return sb.ToString();
         }
         #endregion
     }
